+++ samples/show_conf_signature.ml
@@ -1,0 +1,3 @@
+#use "topfind";;
+#require "dkml-c-probe";;
+#show Dkml_c_probe.C_conf;;